+++ prov/tests/test_json.py
@@ -1,18 +1,12 @@
-"""
-Created on July 14, 2014
-
-@author: Trung Dong Huynh
-"""
-import logging
-from prov.tests.utility import BaseTestCase
-
 from prov.model import ProvDocument
+from prov.tests.utility import BaseTestCase, RoundTripTestCase
+from prov.tests.test_model import AllTestsBase
 
-
+import logging
 logger = logging.getLogger(__name__)
 
 
-class TestDefaultSerializer(BaseTestCase):
+class TestJSONSerializer(BaseTestCase):
     def test_decoding_unicode_value(self):
         unicode_char = u'\u2019'
         json_content = u'''{
@@ -26,6 +20,10 @@ class TestDefaultSerializer(BaseTestCase):
     }
 }''' % unicode_char
 
-        prov_doc = ProvDocument.deserialize(content=json_content)
+        prov_doc = ProvDocument.deserialize(content=json_content, format='json')
         e1 = prov_doc.get_record('ex:unicode_char')[0]
         self.assertIn(unicode_char, e1.get_attribute('prov:label'))
+
+
+class RoundTripJSONTests(RoundTripTestCase, AllTestsBase):
+    FORMAT = 'json'