+++ prov/tests/utility.py
@@ -15,6 +15,24 @@ logger = logging.getLogger(__name__)
 
 
 class BaseTestCase(unittest.TestCase):
+    def assertIsInstance(self, obj, cls, msg=None):
+        """Python < v2.7 compatibility.  Assert isinstance(obj, cls)"""
+        try:
+            f = super(BaseTestCase, self).assertIsInstance
+        except AttributeError:
+            self.assertTrue(isinstance(obj, cls), msg)
+        else:
+            f(obj, cls, msg)
+
+    def assertIsNotNone(self, obj, *args, **kwargs):
+        """Python < v2.7 compatibility.  Assert 'a' in 'b'"""
+        try:
+            f = super(BaseTestCase, self).assertIsNotNone
+        except AttributeError:
+            self.assertTrue(obj is not None, *args, **kwargs)
+        else:
+            f(obj, *args, **kwargs)
+
     def assertIn(self, a, b, *args, **kwargs):
         """Python < v2.7 compatibility.  Assert 'a' in 'b'"""
         try:
@@ -43,44 +61,17 @@ class BaseTestCase(unittest.TestCase):
             f(a, b, *args, **kwargs)
 
 
-class BaseRoundTripTest(unittest.TestCase):
-    def setUp(self):
-        # a dictionary to hold test documents
-        self._documents = dict()
+class RoundTripTestCase(BaseTestCase):
+    """A serializer test should subclass this class and set the class property FORMAT to the correct value (e.g.
+    'json', 'xml', 'rdf').
+    """
+    FORMAT = None  # a subclass should change this
 
-    def write(self, document, fp):
-        pass
+    def assertRoundTripEquivalence(self, prov_doc, msg=None):
+        if self.FORMAT is None:
+            # This is a dummy test, just return
+            return
 
-    def read(self, fp):
-        return None
-
-    def compare_documents(self, doc_1, doc_2):
-        # Self equality check
-        self.assertEqual(doc_1, doc_1)
-        self.assertEqual(doc_2, doc_2)
-        # PROV-N output
-        logger.debug(doc_1.get_provn())
-        # Equality check
-        try:
-            self.assertEqual(doc_1, doc_2)
-        except AssertionError, e:
-            logger.info(u'---- Document 1 ----\n' + doc_1.get_provn())
-            logger.info(u'---- Document 2 ----\n' + doc_2.get_provn())
-            # Re-raise the exception
-            raise e
-
-    def run_roundtrip_test_document(self, document):
-        stream = StringIO()
-        self.write(document, stream)
-        stream.seek(0)
-        doc_2 = self.read(stream)
-        self.compare_documents(document, doc_2)
-
-
-class ProvJSONRoundTripTest(BaseRoundTripTest):
-    def write(self, document, fp):
-        document.serialize(fp)
-
-    def read(self, fp):
-        return ProvDocument.deserialize(fp)
-
+        serialized_content = prov_doc.serialize(format=self.FORMAT, indent=4)
+        prov_doc_new = ProvDocument.deserialize(content=serialized_content, format=self.FORMAT)
+        self.assertEqual(prov_doc, prov_doc_new, msg)